app/requirements.txt
#requirements
flask
celery
redis
prometheus-flask-exporter
kafka-python
